feat/trees: instead of sampling, serve non-served trees and keep track of it
The 'sample' query parameter for the tree view disappears, and we
replace it with a list view: 'serve_long_unserved_choice'.

This view returns an empty list if the provided filters reduce the
queryset to an empty set, otherwise it orders eligible trees by the time
at which they were last served (by this same view only), and selects one
among the first half of those.

So if you continuously select a tree through this route (as is expected
from the client), two occurrences of the same tree will always be
separated by at least half the number of available trees, while still
keeping some randomness in the sampling.